# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  ageRange?: string | null;
  /** @nullable */
  country?: string | null;
  languageCode: string;
  /** @nullable */
  theme?: string | null;
  createdAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  ageRange?: string;
  country?: string;
  languageCode?: string;
  theme?: string;
}

export type AssessmentResultCategory =
  (typeof AssessmentResultCategory)[keyof typeof AssessmentResultCategory];

export const AssessmentResultCategory = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Assessment {
  id: number;
  userId: string;
  score: number;
  resultCategory: AssessmentResultCategory;
  answers?: string;
  createdAt: string;
}

export interface SubmitAssessmentBody {
  answers: number[];
}

export interface Resource {
  id: number;
  slug: string;
  title: string;
  content: string;
  category: string;
  isActive: boolean;
  createdAt: string;
}

export interface AdminCreateResourceBody {
  slug: string;
  title: string;
  content: string;
  category: string;
  isActive?: boolean;
}

export interface ContactForm {
  id: number;
  /** @nullable */
  userId?: string | null;
  email: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface SubmitContactFormBody {
  email: string;
  message: string;
}

export type HelpRequestUrgencyLevel =
  (typeof HelpRequestUrgencyLevel)[keyof typeof HelpRequestUrgencyLevel];

export const HelpRequestUrgencyLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type HelpRequestStatus =
  (typeof HelpRequestStatus)[keyof typeof HelpRequestStatus];

export const HelpRequestStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface HelpRequest {
  id: number;
  /** @nullable */
  userId?: string | null;
  email: string;
  urgencyLevel: HelpRequestUrgencyLevel;
  message: string;
  status: HelpRequestStatus;
  createdAt: string;
}

export type SubmitHelpRequestBodyUrgencyLevel =
  (typeof SubmitHelpRequestBodyUrgencyLevel)[keyof typeof SubmitHelpRequestBodyUrgencyLevel];

export const SubmitHelpRequestBodyUrgencyLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface SubmitHelpRequestBody {
  email: string;
  urgencyLevel: SubmitHelpRequestBodyUrgencyLevel;
  message: string;
}

export interface AdminStats {
  totalUsers: number;
  totalAssessments: number;
  totalChatConversations: number;
  totalHelpRequests: number;
  newHelpRequests: number;
  activationRate: number;
  assessmentCompletionRate: number;
  helpRequestRate: number;
}

export interface AdminUser {
  userId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  role: string;
  isBlocked: boolean;
  createdAt: string;
  /** @nullable */
  lastAssessmentCategory?: string | null;
}

export interface AdminUsersResponse {
  users: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export type AdminUpdateUserBodyRole =
  (typeof AdminUpdateUserBodyRole)[keyof typeof AdminUpdateUserBodyRole];

export const AdminUpdateUserBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AdminUpdateUserBody {
  isBlocked?: boolean;
  role?: AdminUpdateUserBodyRole;
}

export type AdminUpdateHelpRequestBodyStatus =
  (typeof AdminUpdateHelpRequestBodyStatus)[keyof typeof AdminUpdateHelpRequestBodyStatus];

export const AdminUpdateHelpRequestBodyStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface AdminUpdateHelpRequestBody {
  status: AdminUpdateHelpRequestBodyStatus;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type GetResourcesParams = {
  category?: string;
  locale?: string;
};

export type AdminListUsersParams = {
  page?: number;
  limit?: number;
};

export type AdminListContactFormsParams = {
  status?: string;
};

export type AdminListHelpRequestsParams = {
  status?: string;
};