# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sereni API - Emotional support platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessment:
    post:
      operationId: submitAssessment
      tags: [assessment]
      summary: Submit a new self-assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssessmentBody"
      responses:
        "201":
          description: Assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessment/latest:
    get:
      operationId: getLatestAssessment
      tags: [assessment]
      summary: Get the most recent assessment for the current user
      responses:
        "200":
          description: Latest assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: No assessment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /contact:
    post:
      operationId: submitContactForm
      tags: [contact]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContactFormBody"
      responses:
        "201":
          description: Contact form submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactForm"
  /help-request:
    post:
      operationId: submitHelpRequest
      tags: [contact]
      summary: Submit a help request (professional referral)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitHelpRequestBody"
      responses:
        "201":
          description: Help request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/users/{userId}:
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user (block/unblock, set role)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/contact-forms:
    get:
      operationId: adminListContactForms
      tags: [admin]
      summary: List all contact form submissions
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contact forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactForm"
  /admin/help-requests:
    get:
      operationId: adminListHelpRequests
      tags: [admin]
      summary: List all help requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of help requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HelpRequest"
  /admin/help-requests/{id}:
    put:
      operationId: adminUpdateHelpRequest
      tags: [admin]
      summary: Update help request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateHelpRequestBody"
      responses:
        "200":
          description: Updated help request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
  /admin/resources:
    get:
      operationId: adminListResources
      tags: [admin]
      summary: List all resources (admin view)
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: adminCreateResource
      tags: [admin]
      summary: Create a new resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateResourceBody"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /admin/resources/{id}:
    put:
      operationId: adminUpdateResource
      tags: [admin]
      summary: Update a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateResourceBody"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    delete:
      operationId: adminDeleteResource
      tags: [admin]
      summary: Delete a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Assessment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        score:
          type: integer
        resultCategory:
          type: string
          enum: [low, medium, high]
        answers:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - score
        - resultCategory
        - createdAt
    SubmitAssessmentBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: integer
      required:
        - answers
    Resource:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - content
        - category
        - isActive
        - createdAt
    AdminCreateResourceBody:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isActive:
          type: boolean
      required:
        - slug
        - title
        - content
        - category
    ContactForm:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        email:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - message
        - status
        - createdAt
    SubmitContactFormBody:
      type: object
      properties:
        email:
          type: string
        message:
          type: string
      required:
        - email
        - message
    HelpRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        email:
          type: string
        urgencyLevel:
          type: string
          enum: [low, medium, high]
        message:
          type: string
        status:
          type: string
          enum: [new, contacted, closed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - urgencyLevel
        - message
        - status
        - createdAt
    SubmitHelpRequestBody:
      type: object
      properties:
        email:
          type: string
        urgencyLevel:
          type: string
          enum: [low, medium, high]
        message:
          type: string
      required:
        - email
        - urgencyLevel
        - message
...
    AdminUser:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastAssessmentCategory:
          type: ["string", "null"]
      required:
        - userId
        - email
        - role
        - isBlocked
        - createdAt
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    AdminUpdateUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
        role:
          type: string
          enum: [user, admin]
    AdminUpdateHelpRequestBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, closed]
      required:
        - status
...