# lib/api-zod/src/generated/types/adminCreateResourceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCreateResourceBody {
  slug: string;
  title: string;
  content: string;
  category: string;
  isActive?: boolean;
}