# lib/api-zod/src/generated/types/adminUpdateHelpReuestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUpdateHelpRequestBodyStatus } from "./adminUpdateHelpRequestBodyStatus";

export interface AdminUpdateHelpRequestBody {
  status: AdminUpdateHelpRequestBodyStatus;
}