# lib/api-zod/src/generated/types/contactForm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */

export interface ContactForm {
  id: number;
  /** @nullable */
  userId?: string | null;
  email: string;
  message: string;
  status: string;
  createdAt: Date;
}