# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  slug: string;
  title: string;
  content: string;
  category: string;
  isActive: boolean;
  createdAt: Date;
}