# lib/api-zod/src/generated/types/submitContactFormBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitContactFormBody {
  email: string;
  message: string;
}